remove FeatureFlag const classes ExcludedSectionListProvider moved to LtiProvider directory FeatureFlagFormPropertyMapper using `$_ENV` variable
